refactor(core): rename `Managed*` section data files
None of the shstrtab nor garbage collection stuff was limited to managed->packed conversion, so I've opted to rename things.

Have everything be named some variant of `managed_section_*`
