+++ core/core/formats/elf/managed_section_strtab.hpp
